requirements.txt
asgiref==3.8.1
Django==5.1.2
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
djangorestframework==3.15.2
pillow==11.0.0
sqlparse==0.5.1
tzdata==2024.2
